Update models and firebase implementation
