+++ domain/src/main/kotlin/team/dahaeng/android/domain/schedule/model/Schedule.kt
@@ -9,11 +9,11 @@
 
 package team.dahaeng.android.domain.schedule.model
 
-import team.dahaeng.android.domain.schedule.model.travel.Travel
-import team.dahaeng.android.domain.util.constant.EmptyString
 import java.io.Serializable
 import java.util.Date
 import kotlin.random.Random
+import team.dahaeng.android.domain.schedule.model.travel.Travel
+import team.dahaeng.android.domain.util.constant.EmptyString
 
 /**
  * 일정 객체
@@ -25,20 +25,24 @@ import kotlin.random.Random
  * 이 아이템을 이용해 본인만의 일정으로 새로 작성하게 된다면 (fork 와 유사)
  * copy 로 [id] 만 바꿔서 재사용 하면 됨
  *
+ * @property createdAt 일정이 업로드된 시간
  * @property id 일정 UUID
  * @property ownerId 일정 등록자 UUID
  * @property participant 일정 참여자 UUID 리스트
  * @property title 일정 제목
  * @property content 일정 내용, HTML 컨텐츠 텍스트 span 한정으로 지원
  * @property travel 일정에 첨부된 여행 정보
- * @property createdAt 일정이 업로드된 시간
+ * @property forkCount 일정이 쓰인 휫수 (fork된 휫수)
+ * @property bookmarkedCount 일정이 북마크된 휫수
  */
 data class Schedule(
+    val createdAt: Long = Date().time,
     val id: Long = Random.nextLong(),
     val ownerId: Long = 0L,
     val participant: List<Long> = emptyList(),
     val title: String = EmptyString,
     val content: String = EmptyString,
     val travel: Travel = Travel(),
-    val createdAt: Long = Date().time,
+    val forkCount: Int = 0,
+    val bookmarkedCount: Int = 0,
 ) : Serializable
